re refactor Services-Interfaces
